requirements.txt
# Core
langchain>=0.3.0
langchain-anthropic>=0.3.0
langchain-core>=0.3.0
langgraph>=0.2.0

# LLM provider
anthropic>=0.40.0

# Output integrations
notion-client>=2.2.0

# Data validation
pydantic>=2.0.0

# Environment
python-dotenv>=1.0.0
